Create .gitignore, add IDE metadata to .gitignore, begin proper README
The project had no .gitignore because we recently  moved it to git. The
most popular format for the README is markdown, so I reformatted the file.

The info contained in the readme is way out of date and needs to be updated,
but I have to commit the .gitignore before I create a project in my IDE.